feat: 강의개설API 구현중

+++ core-service/src/main/java/com/green/core/application/major/MajorRepository.java
@@ -3,5 +3,5 @@ package com.green.core.application.major;
 import com.green.core.entity.major.Major;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MajorRepository extends JpaRepository<Major, Integer> {
+public interface MajorRepository extends JpaRepository<Major, Long> {
 }
